Revert "Pedro/bal 431 handle metastable conversions inside math package (#119)"
This reverts commit c81cf73a5b52644b37fb69ede027bec9bf009ff9.
